Include clean function in JS API
